modify "rmmap" tests for SLE15
On older SLE kernels, it was possible to add maps even with
zero available paths. On SLE15, this is still possible for map
reload, but not for creation of new maps.

Change the tests to not fail always.

CAUTION: "remove map" tests always fail if anything but "raw"
partitions are used.